backend/requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sentence-transformers==2.4.0
faiss-cpu==1.7.4
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
pytest==7.4.4
httpx==0.27.0
